[Refactor] Service에서 트랜잭션 분리
